Rewrote renderer and scene class.
