+++ Core/Core.vcxproj
@@ -105,7 +105,6 @@
     <ClCompile Include="imgui\imgui_impl_opengl3.cpp" />
     <ClCompile Include="imgui\imgui_tables.cpp" />
     <ClCompile Include="imgui\imgui_widgets.cpp" />
-    <ClCompile Include="Renderer.cpp" />
   </ItemGroup>
   <ItemGroup>
     <CudaCompile Include="RayTracer.cu">
@@ -122,7 +121,6 @@
     <ClInclude Include="CUDAOutputBuffer.h" />
     <ClInclude Include="Exception.h" />
     <ClInclude Include="RayTracer.h" />
-    <ClInclude Include="hitable_list.h" />
     <ClInclude Include="imgui\imconfig.h" />
     <ClInclude Include="imgui\imgui.h" />
     <ClInclude Include="imgui\imgui_impl_glfw.h" />
