+++ Core/material.h
@@ -13,8 +13,15 @@ public:
 		Dielectric
 	};
 
+	Material(Type type, const glm::vec3& albedo = glm::vec3(0.0f), float fuzz = 0, float ir = 0)
+		: m_type(type), m_albedo(albedo), m_fuzz(fuzz), m_ir(ir)
+	{
+	}
+
+
+
 	inline const Type GetType() const { return m_type; }
-	inline const float3& GetAlbedo() const { return make_float3(m_albedo.r, m_albedo.g, m_albedo.b); }
+	inline float3 GetAlbedo() const { return make_float3(m_albedo.r, m_albedo.g, m_albedo.b); }
 	inline const float GetFuzz() const { return m_fuzz; }
 	inline const float GetIR() const { return m_ir; }
